Update Postman collection
Changed the tests so that they rely on a published Postman collection,
accessible here: https://github.com/gothinkster/realworld/tree/master/api
Besides the fact that it is up-to-date, the script is written in a way
that randomises the username, so it can successfully run multiple times
in a row.

A few changes had to performed to support this change:
- The tests expect User to always be returned with bio and image, so
they are set to an empty string by default.
- The tests expect Comment to always be returned with updateAt, so it is
now provided.